% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_tergm_fit.R
\docType{data}
\name{ego_tergm_fit}
\alias{ego_tergm_fit}
\title{Example Ego-TERGM Output}
\format{Ego-TERGM example output, estimated from the Knecht vignette \code{"network"}}
\usage{
data("ego_tergm_fit")
}
\description{
This is the output from a call to the ego-TERGM estimated according to the Knecht vignette.
}
\keyword{ego-TERGM}
\keyword{fit}
