% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_ego_tergm.R
\name{interpret_ego_tergm}
\alias{interpret_ego_tergm}
\title{Custom interpret function for ego-TERGM.}
\usage{
interpret_ego_tergm(ego_tergm_fit = NULL, custom_var_names = NULL)
}
\arguments{
\item{ego_tergm_fit}{The output from a fitted "ego_tergm".}

\item{custom_var_names}{A vector of character terms in the same order as the form object fed to ego_tergm of clearer names for these variables.}
}
\description{
This function assists the user in interpreting ouput from the ego_tergm function.
}
\examples{
interpret_ego_tergm(ego_tergm_fit = ego_tergm_fit)
interpret_ego_tergm(ego_tergm_fit = ego_tergm_fit,
                    custom_var_names = c("Edges", "Mutual", "Triangle",
                                        "In-Degree", "Out-Degree", "Sex Homophily"))
}
\keyword{#@references}
\keyword{2017.}
\keyword{Benjamin}
\keyword{Campbell,}
\keyword{Ego-TERGM.}
\keyword{Inferring}
\keyword{Latent}
\keyword{Longitudinal}
\keyword{Networks}
\keyword{Paper.}
\keyword{Roles}
\keyword{W.}
\keyword{Working}
\keyword{in}
\keyword{interpretation}
\keyword{summary}
\keyword{the}
\keyword{using}
