% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_egonets.R
\name{sim_egonets}
\alias{sim_egonets}
\title{Simulation of longitudinal ego-networks according to a mixture of data generating processes}
\usage{
sim_egonets(form = NULL, params = NULL, roles = NULL, N_per_role = NULL,
  t_steps = NULL, egonet_size = NULL)
}
\arguments{
\item{form}{A vector of ERGM terms that should be used to generate the networks.}

\item{params}{A "roles" by "form" matrix of ERGM simulation parameters with groups defined on the row and model terms defined on the column.}

\item{roles}{An integer for the number of distinct mixture groups that should be simulated.}

\item{N_per_role}{An integer for the number of different longitudinally observed ego-networks that should be simulated per role in roles.}

\item{t_steps}{An integer for the number of time steps that each ego-network should be observed across.}

\item{egonet_size}{An integer for the size of each ego-network simulated.}
}
\value{
A list of simulated ego-networks.
}
\description{
This function estimates an ego-TERGM.
}
\examples{
net <- sim_egonets(form = c("edges", "gwesp(0.8,fixed=TRUE)", "gwdegree(decay=0.8,fixed=TRUE)"),
                  params = rbind(c(-3,1,0), c(-1,-2,-1), c(-2,0,2)),
                  roles = 3,
                  N_per_role = 10,
                  t_steps = 3,
                  egonet_size = 20)
}
\references{
Campbell, Benjamin W. 2017. Inferring Latent Roles in Longitudinal Networks using the Ego-TERGM. Working Paper.
}
\keyword{networks}
\keyword{simulation}
